Make `values` take an `Iterable` of non-null values
Unless further constraints have been defined for a type, a type that takes in just `T`, when erased to `*`, will be inferred as `Any?`, which I don't think is the intended behavior here.

It doesn't help that Kotlin didn't complain about it either because `Any?.toString()` is a valid function, which will just returns `"null"` on `null` value.